requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# AI/ML dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
langgraph==0.0.20
openai==1.3.0

# Google Calendar integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
zoneinfo==0.2.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

